--- pages/page2.py
@@ -1,10 +1,0 @@
-import dash
-from dash.dependencies import Input, Output, State
-from dash import dash_table, dcc, html
-
-import pandas as pd
-import plotly.express as px
-
-layout = html.Div([
-    html.H3("Page 2")
-])
